requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.12.0
blinker==1.5
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
debugpy==1.6.4
decorator==5.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-crontab==0.7.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
dnspython==2.2.1
drf-yasg==1.21.4
email-validator==1.3.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
filelock==3.8.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
ipykernel==6.19.4
ipython==8.7.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.1
kaitaistruct==0.10
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.24.1
openapi-codec==1.3.2
outcome==1.2.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pathspec==0.10.3
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
py3-validate-email==1.0.5.post1
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydivert==2.1.0
Pygments==2.13.0
pyIsEmail==2.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.7
pywin32==305
pyzmq==24.0.1
requests==2.28.1
requests-file==1.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.5.0
selenium-requests==2.0.2
selenium-wire==5.1.0
shortuuid==1.0.11
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
tldextract==3.4.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
undetected-chromedriver==3.2.1
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
webdriver-manager==3.8.4
websockets==10.4
wsproto==1.2.0
zstandard==0.19.0
